automatically collect and setup all instantiated loggers
